\name{CALDER-package}
\alias{CALDER-package}
\alias{CALDER}
\docType{package}
\title{
\packageTitle{CALDER}
}
\description{
\packageDescription{CALDER}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CALDER}
\packageIndices{CALDER}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CALDER}

Maintainer: \packageMaintainer{CALDER}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
